Added shuffle all feature to library
Feature is a hybrid of serverside and clientside shuffle solutions to help offset the detrimental effects of large audio libraries being held in clientside javascript reserved memory.

Also some general cleanup, and a general moving of everything in the audio extension to AJAX to keep navigating around from interfering with audio playback.