Added Media Ext Settings
Added settings form for Media Ext, updated files to reflect settings, and updated framework to fix issues with put_var and add support for setting a var as session with the put_var function.

Downgraded error reporting to E_ERROR to avoid "no file" notices in the settings module. This will be a sever setting at a later date.