Video Player Mobile Compat
Updated video player to work better with mobile. (Auto requesting landscape was included in a previous update, but was ineffective as the fullscreen toggle control was not visible in portrait mode)

Added logic that if the video is within one minute of being over (as loaded by the history), we just load the video at the beginning.