Developer should have been better about updating update,
Installer system was broken for a while. This update fixes the installer

Additionally, this update fixes a number of broken ajax calls for the playlist system as well as implementing correct renaming and deleting behavior for playlists.

Vogon core was recently reworked to better support POST REDIRECT and GET best practice for form processing. Updated the Media and Audio extensions to use the new .post controllers to handle edits to media server entries. Users will get an update later on, but needs a more substantial rewrite as the base extension was written many versions of vogon ago.

Added a link to the album from the audio item edit page.

Added df -h and uptime outputs to the server page, as well as an output of the appname and version constants.

misc. bugfixes