Visualizer Update
Reigned in audio samples for the visualizer from the 1024 to 768 because sampling a bunch of zeroes is less interesting and most tracks in testing don't do much with those higher values. Also moved to using a math equation at init to set this value, so further tuning will just be changing the percentage variable used to calculate this.

Cleaned up some additional visualizer code to better size cleanCircle and get more of a background color pulse in all visualizers. Hopefully it's not too far and won't trigger any adverse effects, will have to research that more to see if any further clamps need to be placed on the effects for safety reasons, maybe give the user options to turn certain effects on or off.

TODO: Add a null visualizer for those who would rather have nothing there.