Foundation for update system
Updating the install extension and laying foundational work for update checks and automated update system.

Added ver file to declare the version of the framework before install occurs so it can be compared to other versions.

Added vParse class which is responsible for parse version numbers and determining if a version is greater, lesser, or equal to another.