Import Hotfix
Fixed issue with import script passing array by reference to generate the name for the created screenshot.

Also patched issue where undefined playtime_seconds metadata wouldn't assign 0.

Updated minidlna settings and controls to reflect situations where the service will be run system-wide (this use case requires adding www-data to the users that can sudo without a password.) We'll need to look into this further to see if we can make this more secure in the future.